requirements.txt
alembic==1.18.4
annotated-types==0.7.0
Authlib==1.6.9
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.71.2
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
pillow==12.1.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dotenv==1.2.2
requests==2.32.5
rsa==4.9.1
SQLAlchemy==2.0.48
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.6
WTForms==3.2.1
pgvector==0.3.6
psycopg2-binary==2.9.11
PyPDF2==3.0.1
groq
openai
gunicorn==21.2.0
